Changes to SeLion gecko driver support to align with Se3 assumptions
- default to marionette true
- pay attention to webdriver.firefox.marionette
- provide path to boot local node with legacy ff support
- delete GeckoCapabilitiesBuilder --> move functionality
  to FirefoxCapabilitiesBuilder
- rename TestCapabilitiyBuilder which is specific to Travis
- update ci scripts
