+++ client/src/main/java/com/paypal/selion/internal/platform/grid/browsercapabilities/FireFoxCapabilitiesBuilder.java
@@ -1,5 +1,5 @@
 /*-------------------------------------------------------------------------------------------------------------------*\
-|  Copyright (C) 2014 PayPal                                                                                          |
+|  Copyright (C) 2014-2016 PayPal                                                                                     |
 |                                                                                                                     |
 |  Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance     |
 |  with the License.                                                                                                  |
@@ -24,6 +24,8 @@ import org.openqa.selenium.firefox.internal.ProfilesIni;
 import org.openqa.selenium.remote.CapabilityType;
 import org.openqa.selenium.remote.DesiredCapabilities;
 
+import com.paypal.selion.SeLionConstants;
+import com.paypal.selion.configuration.Config;
 import com.paypal.selion.configuration.Config.ConfigProperty;
 import com.paypal.selion.platform.grid.browsercapabilities.DefaultCapabilitiesBuilder;
 
@@ -39,6 +41,14 @@ class FireFoxCapabilitiesBuilder extends DefaultCapabilitiesBuilder {
         capabilities.setCapability(FirefoxDriver.PROFILE, prepareFireFoxProfile());
         if (ProxyHelper.isProxyServerRequired()) {
             capabilities.setCapability(CapabilityType.PROXY, ProxyHelper.createProxyObject());
+        }
+
+        if (!isLegacyFF()) {
+            capabilities.setCapability(FirefoxDriver.MARIONETTE, true);
+            String geckoDriverPath = getBinaryPath();
+            if (isLocalRun() && StringUtils.isNotBlank(geckoDriverPath)) {
+                System.setProperty(SeLionConstants.WEBDRIVER_GECKO_DRIVER_PROPERTY, geckoDriverPath);
+            }
         }
 
         return capabilities;
@@ -111,4 +121,18 @@ class FireFoxCapabilitiesBuilder extends DefaultCapabilitiesBuilder {
         logger.exiting(profile);
         return profile;
     }
+
+    /*
+     * Returns the location of Gecko/marionette driver or empty string if it cannot be determined.
+     */
+    private String getBinaryPath() {
+        String location = System.getProperty(SeLionConstants.WEBDRIVER_GECKO_DRIVER_PROPERTY,
+                Config.getConfigProperty(ConfigProperty.SELENIUM_GECKODRIVER_PATH));
+        return location;
+    }
+
+    private boolean isLegacyFF() {
+        return !Config.getBoolConfigProperty(ConfigProperty.SELENIUM_USE_GECKODRIVER);
+    }
+
 }
